// src/signalr-types.ts
export enum EqualizerStreamData {
  selfTestResult = 1,
  selfTestDetails = 2,
  easeeLinkCommandResponse = 13,
  easeeLinkDataReceived = 14,
  siteIDNumeric = 19,
  config_siteStructure = 20,
  state_softwareRelease = 21,
  config_meterType = 25,
  config_meterID = 26,
  oBISListIdentifier = 27,
  config_gridType = 29,
  config_numPhases = 30,
  state_currentL1 = 31,
  state_currentL2 = 32,
  state_currentL3 = 33,
  state_voltageNL1 = 34,
  state_voltageNL2 = 35,
  state_voltageNL3 = 36,
  state_voltageL1L2 = 37,
  state_voltageL1L3 = 38,
  state_voltageL2L3 = 39,
  state_activePowerImport = 40,
  state_activePowerExport = 41,
  state_reactivePowerImport = 42,
  state_reactivePowerExport = 43,
  state_maxPowerImport = 44,
  state_cumulativeActivePowerImport = 45,
  state_cumulativeActivePowerExport = 46,
  state_cumulativeReactivePowerImport = 47,
  state_cumulativeReactivePowerExport = 48,
  state_clockAndDateMeter = 49,
  state_rcpi = 50,
  config_ssid = 51,
  config_masterBackPlateID = 55,
  config_equalizerID = 56,
  exceptionData = 60,
  bootReason = 61,
  highCurrentTransitions = 64,
  state_vCap = 65,
  state_vBusMin = 66,
  state_vbusMax = 67,
  state_internalTemperature = 68,
  hanSnapshot = 69,
  state_localRSSI = 70,
  localTxPower = 71,
  localRadioChannel = 72,
  localShortAddress = 73,
  localNodeType = 74,
  localParentAddress = 75,
  circuitPhaseMapping = 80,
  phaseMappingReport = 81,
  modbusConfiguration = 85,
  state_loadbalanceThrottle = 86,
  state_availableCurrentL1 = 87,
  state_availableCurrentL2 = 88,
  state_availableCurrentL3 = 89,
  hanChecksumErrors = 90,
  APMacAddress = 91,
  wifiReconnects = 92,
  state_ledMode = 100,
}